+++ src/internal/streamstore/details_test.go
@@ -7,6 +7,7 @@ import (
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 
+	"github.com/alcionai/clues"
 	"github.com/alcionai/corso/src/internal/kopia"
 	"github.com/alcionai/corso/src/internal/tester"
 	"github.com/alcionai/corso/src/pkg/backup/details"
@@ -27,32 +28,34 @@ func TestStreamDetailsIntegrationSuite(t *testing.T) {
 }
 
 func (suite *StreamDetailsIntegrationSuite) TestDetails() {
-	t := suite.T()
-
 	ctx, flush := tester.NewContext()
 	defer flush()
 
+	t := suite.T()
 	// need to initialize the repository before we can test connecting to it.
 	st := tester.NewPrefixedS3Storage(t)
-
 	k := kopia.NewConn(st)
-	require.NoError(t, k.Initialize(ctx))
+
+	err := k.Initialize(ctx)
+	require.NoError(t, err, clues.ToCore(err))
 
 	defer k.Close(ctx)
 
 	kw, err := kopia.NewWrapper(k)
-	require.NoError(t, err)
+	require.NoError(t, err, clues.ToCore(err))
 
 	defer kw.Close(ctx)
 
 	deetsBuilder := &details.Builder{}
 
-	deetsBuilder.Add("ref", "shortref", "parentref", "locationRef", true,
-		details.ItemInfo{
-			Exchange: &details.ExchangeInfo{
-				Subject: "hello world",
-			},
-		})
+	require.NoError(
+		t,
+		deetsBuilder.Add("ref", "shortref", "parentref", "locationRef", true,
+			details.ItemInfo{
+				Exchange: &details.ExchangeInfo{
+					Subject: "hello world",
+				},
+			}))
 
 	var (
 		deets = deetsBuilder.Details()
@@ -60,12 +63,12 @@ func (suite *StreamDetailsIntegrationSuite) TestDetails() {
 	)
 
 	id, err := sd.Write(ctx, deets, fault.New(true))
-	require.NoError(t, err)
+	require.NoError(t, err, clues.ToCore(err))
 	require.NotNil(t, id)
 
 	var readDeets details.Details
 	err = sd.Read(ctx, id, details.UnmarshalTo(&readDeets), fault.New(true))
-	require.NoError(t, err)
+	require.NoError(t, err, clues.ToCore(err))
 	require.NotEmpty(t, readDeets)
 
 	assert.Equal(t, len(deets.Entries), len(readDeets.Entries))