+++ src/pkg/account/account_test.go
@@ -3,6 +3,7 @@ package account
 import (
 	"testing"
 
+	"github.com/alcionai/clues"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
 )
@@ -44,7 +45,7 @@ func (suite *AccountSuite) TestNewAccount() {
 	for _, test := range table {
 		suite.T().Run(test.name, func(t *testing.T) {
 			s, err := NewAccount(test.p, test.c)
-			test.errCheck(t, err)
+			test.errCheck(t, err, clues.ToCore(err))
 			// remaining tests are dependent upon error-free state
 			if test.c.err != nil {
 				return