+++ src/pkg/backup/details/details.go
@@ -152,10 +152,11 @@ func (b *Builder) Add(
 	repoRef, shortRef, parentRef, locationRef string,
 	updated bool,
 	info ItemInfo,
-) {
+) error {
 	b.mu.Lock()
 	defer b.mu.Unlock()
-	b.d.add(repoRef, shortRef, parentRef, locationRef, updated, info)
+
+	return b.d.add(repoRef, shortRef, parentRef, locationRef, updated, info)
 }
 
 func (b *Builder) Details() *Details {
@@ -283,15 +284,52 @@ func (d *Details) add(
 	repoRef, shortRef, parentRef, locationRef string,
 	updated bool,
 	info ItemInfo,
-) {
-	d.Entries = append(d.Entries, DetailsEntry{
+) error {
+	entry := DetailsEntry{
 		RepoRef:     repoRef,
 		ShortRef:    shortRef,
 		ParentRef:   parentRef,
 		LocationRef: locationRef,
 		Updated:     updated,
 		ItemInfo:    info,
-	})
+	}
+
+	// Use the item name and the path for the ShortRef. This ensures that renames
+	// within a directory generate unique ShortRefs.
+	if info.infoType() == OneDriveItem || info.infoType() == SharePointLibrary {
+		p, err := path.FromDataLayerPath(repoRef, true)
+		if err != nil {
+			return clues.Wrap(err, "munging OneDrive or SharePoint ShortRef")
+		}
+
+		if info.OneDrive == nil && info.SharePoint == nil {
+			return clues.New("item is not SharePoint or OneDrive type")
+		}
+
+		filename := ""
+		if info.OneDrive != nil {
+			filename = info.OneDrive.ItemName
+		} else if info.SharePoint != nil {
+			filename = info.SharePoint.ItemName
+		}
+
+		// Make the new path contain all display names and then the M365 item ID.
+		// This ensures the path will be unique, thus ensuring the ShortRef will be
+		// unique.
+		//
+		// If we appended the file's display name to the path then it's possible
+		// for a folder in the parent directory to have the same display name as the
+		// M365 ID of this file and also have a subfolder in the folder with a
+		// display name that matches the file's display name. That would result in
+		// duplicate ShortRefs, which we can't allow.
+		elements := p.Elements()
+		elements = append(elements[:len(elements)-1], filename, p.Item())
+		entry.ShortRef = path.Builder{}.Append(elements...).ShortRef()
+	}
+
+	d.Entries = append(d.Entries, entry)
+
+	return nil
 }
 
 // addFolder adds an entry for the given folder.